# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task API",
        "description": "API para gestionar tareas.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/operUsers": {
            "get": {
                "tags": [
                    "Operational Users"
                ],
                "summary": "Lista todos los usuarios operativos",
                "operationId": "e9a64bed9d368678d95a463744e44346",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios operativos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Operational Users"
                ],
                "summary": "Almacena un nuevo usuario operativo",
                "operationId": "41ac362a32670733395a0253b91a826a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreOperUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirecciona a la lista de usuarios operativos con mensaje de éxito."
                    }
                }
            }
        },
...
        "/operUsers/{id}": {
            "get": {
                "tags": [
                    "Operational Users"
                ],
                "summary": "Muestra los detalles de un usuario operativo",
                "operationId": "a0ac02f99140c32d3214dc4006e8bac8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario operativo.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del usuario operativo.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Operational Users"
                ],
                "summary": "Actualiza un usuario operativo existente",
                "operationId": "a31b3625252ca689baa4adcdbcf30cb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario operativo a actualizar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModifyOperUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirecciona a la lista de usuarios operativos con mensaje de éxito."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Operational Users"
                ],
                "summary": "Elimina un usuario operativo",
                "operationId": "0fc23662b8504013a3b40e83b06369aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario operativo a eliminar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirecciona a la lista de usuarios operativos con mensaje de éxito."
                    }
                }
            }
        },
        "/operUsers/{user}/edit": {
            "get": {
                "tags": [
                    "Operational Users"
                ],
                "summary": "Muestra el formulario para editar un usuario operativo",
                "operationId": "87a2af0268ee163081f9dab0252b0220",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID del usuario operativo a editar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulario de edición de usuario operativo."
                    }
                }
            }
        },
        "/provinces": {
            "get": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Lista todas las provincias",
                "operationId": "59577da69a733e49ff503cd1100a2860",
                "responses": {
                    "200": {
                        "description": "Lista de provincias.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Province"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Almacena una nueva provincia",
                "operationId": "26afd240838d142bac073ce82c8f4234",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Province"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirecciona a la lista de provincias con mensaje de éxito."
                    }
                }
            }
        },
...
        "/provinces/{province}/edit": {
            "get": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Muestra el formulario para editar una provincia",
                "operationId": "938f77410553890fc09744edca8867fe",
                "parameters": [
                    {
                        "name": "province",
                        "in": "path",
                        "description": "ID de la provincia a editar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulario de edición de provincia."
                    }
                }
            }
        },
        "/provinces/{province}": {
            "put": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Actualiza una provincia existente",
                "operationId": "86ed57588821530ac39e2a2f04fbb238",
                "parameters": [
                    {
                        "name": "province",
                        "in": "path",
                        "description": "ID de la provincia a actualizar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Province"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirecciona a la lista de provincias con mensaje de éxito."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Provinces"
                ],
                "summary": "Elimina una provincia",
                "operationId": "11883907aeeac037ab938af08104f5b3",
                "parameters": [
                    {
                        "name": "province",
                        "in": "path",
                        "description": "ID de la provincia a eliminar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirecciona a la lista de provincias con mensaje de éxito."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ModifyOperUserRequest": {
                "title": "Modify Oper User Request",
                "description": "Request para modificar un usuario operativo existente.",
                "required": [
                    "name",
                    "email",
                    "dni",
                    "type"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del usuario.",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del usuario.",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Contraseña del usuario (opcional).",
                        "type": "string",
                        "format": "password"
                    },
                    "dni": {
                        "description": "DNI del usuario.",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Teléfono del usuario (opcional).",
                        "type": "string"
                    },
                    "address": {
                        "description": "Dirección del usuario (opcional).",
                        "type": "string"
                    },
                    "type": {
                        "description": "Tipo de usuario (oper, admin, client).",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "StoreOperUserRequest": {
                "title": "Store Oper User Request",
                "description": "Request para almacenar un nuevo usuario operativo.",
                "required": [
                    "name",
                    "email",
                    "password",
                    "dni",
                    "phone",
                    "address",
                    "type"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del usuario.",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del usuario.",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Contraseña del usuario.",
                        "type": "string",
                        "format": "password"
                    },
                    "dni": {
                        "description": "DNI del usuario.",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Teléfono del usuario.",
                        "type": "string"
                    },
                    "address": {
                        "description": "Dirección del usuario.",
                        "type": "string"
                    },
                    "type": {
                        "description": "Tipo de usuario (oper).",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Province": {
                "title": "Province",
                "description": "Modelo de provincia.",
                "properties": {
                    "cod": {
                        "description": "Código de la provincia.",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre de la provincia.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "Modelo de usuario.",
                "properties": {
                    "id": {
                        "description": "ID del usuario.",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del usuario.",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del usuario.",
                        "type": "string",
                        "format": "email"
                    },
                    "dni": {
                        "description": "DNI del usuario.",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Teléfono del usuario.",
                        "type": "string"
                    },
                    "address": {
                        "description": "Dirección del usuario.",
                        "type": "string"
                    },
                    "type": {
                        "description": "Tipo de usuario (admin, oper).",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación del usuario.",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}